requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
openai
chromadb
tiktoken
python-dotenv
pdfplumber